fix(api): split routes
